// src/app/shared/store/types/formulation.type.ts
export enum FormulationActionTypes {
  GET_CHEMICAL_FORMULATIONS = 'GET_CHEMICAL_FORMULATIONS',

  FILTER_CHEMICAL_FORMULATIONS = 'FILTER_CHEMICAL_FORMULATIONS',
  FILTER_CHEMICAL_FORMULATIONS_SUCCESS = 'FILTER_CHEMICAL_FORMULATIONS_SUCCESS',
  FILTER_CHEMICAL_FORMULATIONS_FAILURE = 'FILTER_CHEMICAL_FORMULATIONS_FAILURE',

  CLEAR_FILTER_CHEMICAL_FORMULATIONS = 'CLEAR_FILTER_CHEMICAL_FORMULATIONS',

  UPDATE_SELECTED_FORMULATIONS = 'UPDATE_SELECTED_FORMULATIONS',

  UPDATE_FORMULATION_ITEM_DETAILS = 'UPDATE_FORMULATION_ITEM_DETAILS',

  DELETE_FORMULATION_ITEM = 'DELETE_FORMULATION_ITEM',

  UPDATE_FORMULATION_MODEL = 'UPDATE_FORMULATION_MODEL',

  SET_PRODUCT_NAME = 'SET_PRODUCT_NAME',

  GET_QUALITY_CONTROL_TESTS = 'GET_QUALITY_CONTROL_TESTS',

  UPDATE_QUALITY_CONTROL_TESTS = 'UPDATE_QUALITY_CONTROL_TESTS',

  RESET_FORMULATION_STORE = 'RESET_FORMULATION_STORE',

  SET_PRODUCT_TYPE = 'SET_PRODUCT_TYPE',

  CLEAR_STATE = 'CLEAR_STATE',
  GET_DROPDOWN_LIST_SUCCESS_FORMULATION = 'GET_DROPDOWN_LIST_SUCCESS_FORMULATION',
  GET_DROPDOWN_LIST_FAILURE_FORMULATION = 'GET_DROPDOWN_LIST_FAILURE_FORMULATION',
  RESET_RESTRICTION = 'RESET_RESTRICTION'
}